+++ webots_ros2_driver/include/webots_ros2_driver/WebotsNode.hpp
@@ -45,11 +45,14 @@ namespace webots_ros2_driver {
     std::unordered_map<std::string, std::string> getDeviceRosProperties(const std::string &name) const;
     std::unordered_map<std::string, std::string> getPluginProperties(tinyxml2::XMLElement *pluginElement) const;
     void setAnotherNodeParameter(std::string anotherNodeName, std::string parameterName, std::string parameterValue);
+    void replaceUrdfNames(std::string &urdf);
     rclcpp::Client<rcl_interfaces::srv::SetParameters>::SharedPtr mClient;
 
     std::string mRobotDescriptionParam;
     std::string mRobotDescription;
     bool mSetRobotStatePublisher;
+    std::string mComponentsRemappingFilePath;
+    std::unordered_map<std::string, std::string> mComponentsRemapping;
 
     bool mWaitingForUrdfRobotToBeRemoved = false;
     rclcpp::Publisher<std_msgs::msg::String>::SharedPtr mRemoveUrdfRobotPublisher;