+++ webots_ros2_universal_robot/CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog for package webots_ros2_universal_robot
 * Clean simulation reset in launch file.
 * Fixed JTC interpolation.
 * Fixed deprecated load_yaml() function.
+* Update driver node to new WebotsController node.
 
 2023.0.4 (2023-05-23)
 ------------------