+++ webots_ros2_universal_robot/CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog for package webots_ros2_universal_robot
 2023.1.0 (2023-XX-XX)
 ------------------
 * Clean simulation reset in launch file.
+* Fixed JTC interpolation.
+* Fixed deprecated load_yaml() function.
 * Update driver node to new WebotsController node.
 
 2023.0.4 (2023-05-23)