+++ webots_ros2_universal_robot/CHANGELOG.rst
@@ -2,7 +2,7 @@
 Changelog for package webots_ros2_universal_robot
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-2023.1.0 (2023-XX-XX)
+2023.1.0 (2023-06-29)
 ------------------
 * Clean simulation reset in launch file.
 * Fixed JTC interpolation.